Expose main nesting function
It seems to run, but the results are not working yet. I think that the problem is that I'm making copies of the items to convert the Python list to a vector of items. The copy is modified by the nest function, but the originals in the Python list are not modified.

Contributes to issue CURA-7259.