Don't expose the BottomLeftPlacer
You can only use this placer by supplying it as a template parameter to the nester. However template parameters are not possible in the Python syntax or via SIP. So the only way to allow the user to use this class is to create an overload of the nest function or something that uses the BLPlacer instead.

Contributes to issue CURA-7259.