Document the Rectangle class and allow fixing them
This is an item, so all of the behaviour of the item is copied. I've copied most of the documentation.
Only vertex() and setVertex() are not exposed because the Rectangle doesn't support them. Nor should it support that.

I also saw that I had forgotten to expose markAsFixedInBin() and isFixed(). So that is now possible too.

Contributes to issue CURA-7259.